Smith predictor
This is a smith predictor that has been tested on a boost converter with the parameters listed below
L = 860uF
C = 470uH
Vin = 6V
Vout = 12V
R = 10
It shows large overshoot (almost 100%) so it's better not to use it if you don't want to lose your boost converter. Also the algorithm have not been optimized so the sampling frequency is only about 100Hz.