Add files via upload
Conventional IMC controller design are based on eliminating the non-minimal part and then design a controller for the minimal phase part of the original system. This may be useful but is not suitable for system with large overshoot/ long time delay. Based on the structure of Smith predictor, the coder found there may be some improvement on IMC structure by adding an additional branch.